refactor: App 컴포넌트 분리 및 export default 위치 수정
- export default 는 HOC 등 중간에 기능이 추가되는 경우에만 분리해서 써준다면
컴포넌트에 붙여서 사용할 때와 분리해서 사용할 때가 확실히 구분되어서 좋을 것 같다!

- App도 컴포넌트 이므로 index에서 분리
