+++ src/components/Button.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-function Button({ onClick, children }) {
+export default function Button({ onClick, children }) {
   return (
     <button
       type="button"
@@ -10,5 +10,3 @@ function Button({ onClick, children }) {
     </button>
   );
 }
-
-export default Button;
