+++ src/components/Counter.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-function Counter({ count = 0, onClick }) {
+export default function Counter({ count = 0, onClick }) {
   return (
     <button type="button" onClick={onClick}>
       Click me! (
@@ -9,5 +9,3 @@ function Counter({ count = 0, onClick }) {
     </button>
   );
 }
-
-export default Counter;
